[fix] Nullable parameter in conversion operator
