+++ Tests/Model/LinqDataTypes.cs
@@ -136,102 +136,4 @@ namespace Tests.Model
 			return string.Format("{{{0,2}, {1,7}, {2:O}, {3,5}, {4}, {5}, '{6}'}}", ID, MoneyValue, DateTimeValue, BoolValue, GuidValue, SmallIntValue, StringValue);
 		}
 	}
-
-	[Table("LinqDataTypes")]
-	public class LinqDataTypes3
-	{
-		public struct CustomMoneyType : IConvertible
-		{
-			public decimal? Amount;
-
-			public static explicit operator CustomMoneyType(decimal? amount) => new CustomMoneyType{Amount = amount};
-
-			public TypeCode GetTypeCode()
-			{
-				throw new NotImplementedException();
-			}
-
-			public bool ToBoolean(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public char ToChar(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public sbyte ToSByte(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public byte ToByte(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public short ToInt16(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public ushort ToUInt16(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public int ToInt32(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public uint ToUInt32(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public long ToInt64(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public ulong ToUInt64(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public float ToSingle(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public double ToDouble(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public decimal ToDecimal(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public DateTime ToDateTime(IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-
-			public string ToString(IFormatProvider provider)
-			{
-				return Amount != null ? Amount.Value.ToString() : null;
-			}
-
-			public object ToType(Type conversionType, IFormatProvider provider)
-			{
-				throw new NotImplementedException();
-			}
-		}
-
-		[Column(DataType = DataType.Decimal)] public CustomMoneyType MoneyValue;
-	}
 }
